Done ex 4.21

+++ part4/bloglist-backend/controllers/blogs.js
@@ -4,17 +4,32 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const { tokenExtractor } = require('../utils/middleware')
 
-blogsRouter.get('/', async (response) => {
+blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
 blogsRouter.post('/', tokenExtractor, async (request, response) => {
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'Token missing or invalid.' })
+  console.log('TOKEN: ', request.token)
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'Token missing or invalid.' })
+  // }
+
+  let decodedToken
+
+try {
+    decodedToken = jwt.verify(request.token, process.env.SECRET)
+  } catch (error) {
+    console.log('JWT ERROR NAME:', error.name)
+    console.log('JWT ERROR MESSAGE:', error.message)
+    console.log('JWT ERROR:', error)
+    return response.status(401).json({ error: error.message })
   }
+
+  const decode = jwt.decode(request.token)
+  console.log('DECODE: ', decode)
 
   const user = await User.findById(decodedToken.id)
   
@@ -40,7 +55,17 @@ blogsRouter.post('/', tokenExtractor, async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', tokenExtractor, async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'Token missing or invalid.' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'Wrong owner.' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
